requirements.txt
pytest
pyyaml
pydantic
